Main.hs, Definimos el uso de la extensión OverloadedStrings
Explicación de la funcionalidad de esta extensión:

""" String literals, on the other hand, are always of type String, and
are not polymorphic at all. The OverloadedStrings extension corrects
this, making string literals polymorphic over the IsString type class,
which is found in the Data.String module in the base package. That is,
you can write:

a :: String
a = "hello"

b :: Text
b = "hello"

OverloadedStrings also adds IsString to the list of defaultable type
classes, so you can use types like String, Text, and Bytestring in a
default declaration. """

https://www.schoolofhaskell.com/school/to-infinity-and-beyond/pick-of-
the-week/guide-to-ghc-extensions/basic-syntax-extensions